pid_controller/python-zenoh/requirements.txt
eclipse-zenoh==1.3.4
numpy>=1.21.0
matplotlib>=3.5.0
